+++ modules/kogito-maven/3.8.x/added/configure-maven.sh
@@ -37,6 +37,8 @@ function configure() {
     if [ "${SCRIPT_DEBUG}" = "true" ] ; then
         cat "${MAVEN_SETTINGS_PATH}"
     fi
+
+    rm -rf *.bak
 }
 
 # insert settings for HTTP proxy into maven settings.xml if supplied
@@ -80,7 +82,7 @@ function configure_proxy() {
         fi
         xml="$xml\
        </proxy>"
-        sed -i "s|<!-- ### configured http proxy ### -->|${xml}|" "${MAVEN_SETTINGS_PATH}"
+        sed -i.bak "s|<!-- ### configured http proxy ### -->|${xml}|" "${MAVEN_SETTINGS_PATH}"
     fi
 }
 
@@ -92,7 +94,7 @@ function configure_mirrors() {
       <url>$MAVEN_MIRROR_URL</url>\
       <mirrorOf>external:*</mirrorOf>\
     </mirror>"
-        sed -i "s|<!-- ### configured mirrors ### -->|$xml|" "${MAVEN_SETTINGS_PATH}"
+        sed -i.bak "s|<!-- ### configured mirrors ### -->|$xml|" "${MAVEN_SETTINGS_PATH}"
     fi
 }
 
@@ -118,7 +120,7 @@ function ignore_maven_self_signed_certificates() {
 function set_kogito_maven_repo() {
     local kogito_maven_repo_url="${JBOSS_MAVEN_REPO_URL}"
     if [ -n "${kogito_maven_repo_url}" ]; then
-        sed -i "s|https://repository.jboss.org/nexus/content/groups/public/|${kogito_maven_repo_url}|" "${MAVEN_SETTINGS_PATH}"
+        sed -i.bak "s|https://repository.jboss.org/nexus/content/groups/public/|${kogito_maven_repo_url}|" "${MAVEN_SETTINGS_PATH}"
     fi
 }
 
@@ -177,7 +179,7 @@ function _add_maven_repo() {
                     </snapshots>\n\
                 </repository>\n\
                 <!-- ### configured repositories ### -->"
-    sed -i "s|<!-- ### configured repositories ### -->|${repo}|" "${MAVEN_SETTINGS_PATH}"
+    sed -i.bak "s|<!-- ### configured repositories ### -->|${repo}|" "${MAVEN_SETTINGS_PATH}"
 
     local pluginRepo="\n\
                 <pluginRepository>\n\
@@ -198,10 +200,10 @@ function _add_maven_repo() {
                 </pluginRepository>\n\
                 <!-- ### configured plugin repositories ### -->"
 
-    sed -i "s|<!-- ### configured plugin repositories ### -->|${pluginRepo}|" "${MAVEN_SETTINGS_PATH}"
+    sed -i.bak "s|<!-- ### configured plugin repositories ### -->|${pluginRepo}|" "${MAVEN_SETTINGS_PATH}"
 
     # new repo should be skipped by mirror if exists
-    sed -i "s|</mirrorOf>|,!${repo_id}</mirrorOf>|g" "${MAVEN_SETTINGS_PATH}"
+    sed -i.bak "s|</mirrorOf>|,!${repo_id}</mirrorOf>|g" "${MAVEN_SETTINGS_PATH}"
 }
 
 # Finds the environment variable  and returns its value if found.