+++ modules/kogito-swf-builder/added/add-extension.sh
@@ -13,13 +13,9 @@ if [ "${SCRIPT_DEBUG}" = "true" ] ; then
     printenv
 fi
 
-# Call the configure-maven here
-source "${script_dir_path}"/configure-maven.sh
-configure
-
 cd "${PROJECT_ARTIFACT_ID}"
-# auto configure JVM settings
-source "${KOGITO_HOME}"/launch/jvm-settings.sh
+
+source "${script_dir_path}"/configure-jvm-mvn.sh
 
 "${MAVEN_HOME}"/bin/mvn -B ${MAVEN_ARGS_APPEND} \
   -nsu \