+++ modules/kogito-swf-builder/added/build-app.sh
@@ -17,11 +17,9 @@ if [ "${SCRIPT_DEBUG}" = "true" ] ; then
     printenv
 fi
 
-# Call the configure-maven here
-source "${script_dir_path}"/configure-maven.sh
-configure
-
 cd "${PROJECT_ARTIFACT_ID}"
+
+source "${script_dir_path}"/configure-jvm-mvn.sh
 
 if [ ! -z "${QUARKUS_EXTENSIONS}" ]; then
   log_info "Adding extensions '${QUARKUS_EXTENSIONS}'"
@@ -37,9 +35,6 @@ if [ ! -z "${resources_path}" ]; then
 else
   log_warn "-> Nothing to copy from ${resources_path}"
 fi
-
-# auto configure JVM settings
-source "${KOGITO_HOME}"/launch/jvm-settings.sh
 
 "${MAVEN_HOME}"/bin/mvn -B ${MAVEN_ARGS_APPEND} \
   -nsu \