+++ modules/kogito-swf-builder/added/create-app.sh
@@ -12,12 +12,7 @@ if [ "${SCRIPT_DEBUG}" = "true" ] ; then
     printenv
 fi
 
-# Call the configure-maven here
-source "${script_dir_path}"/configure-maven.sh
-configure
-
-# auto configure JVM settings
-source "${KOGITO_HOME}"/launch/jvm-settings.sh
+source "${script_dir_path}"/configure-jvm-mvn.sh
 
 "${MAVEN_HOME}"/bin/mvn -B -s "${MAVEN_SETTINGS_PATH}" \
   -nsu \