+++ scripts/build-quarkus-app.sh
@@ -51,7 +51,7 @@ mvn ${MAVEN_OPTIONS} \
     -DprojectArtifactId="serverless-workflow-project" \
     -DprojectVersionId="1.0.0-SNAPSHOT" \
     -DplatformVersion="${quarkus_platform_version}" \
-    -Dextensions="quarkus-kubernetes,kogito-quarkus-serverless-workflow,kogito-addons-quarkus-knative-eventing" \
+    -Dextensions="quarkus-kubernetes,kogito-quarkus-serverless-workflow,kogito-addons-quarkus-knative-eventing,smallrye-health" \
     io.quarkus.platform:quarkus-maven-plugin:"${quarkus_platform_version}":create
 
 echo "Build quarkus app"
@@ -73,6 +73,7 @@ rm -rfv serverless-workflow-project/src/main/docker
 rm -rfv serverless-workflow-project/.mvn/wrapper
 rm -rfv serverless-workflow-project/mvnw*
 rm -rfv serverless-workflow-project/src/test
+rm -rfv serverless-workflow-project/*.bak
 
 # Maven useless files
 # Needed to avoid Maven to automatically redownload from original Maven repository ...