+++ RELEASE_NOTES.md
@@ -2,7 +2,8 @@
 <!-- Keep them in alphabetical order -->
 ## Enhancements
 
+- [KOGITO-8642](https://issues.redhat.com/browse/KOGITO-8642) - Add Smallrye Health Checks to SW Builder
+
 ## Bug Fixes
 
 ## Known Issues
-